add: test create movie (handler)

+++ internal/movie/repository.go
@@ -12,7 +12,7 @@ type Repository interface {
 	GetAllMoviesByGenre(ctx context.Context, genreID int) ([]domain.Movie, error)
 	GetMovieByID(ctx context.Context, id int) (domain.Movie, error)
 	//GetMovieWithContext(ctx context.Context, id int) (domain.Movie, error)
-	Save(ctx context.Context, b domain.Movie) (int64, error)
+	Save(ctx context.Context, b domain.Movie) (int, error)
 	Exists(ctx context.Context, id int) bool
 	Update(ctx context.Context, b domain.Movie, id int) error
 	Delete(ctx context.Context, id int64) error
@@ -127,7 +127,7 @@ func (r *repository) GetMovieWithContext(ctx context.Context, id int) (domain.Mo
 }
 */
 
-func (r *repository) Save(ctx context.Context, m domain.Movie) (int64, error) {
+func (r *repository) Save(ctx context.Context, m domain.Movie) (int, error) {
 	stm, err := r.db.Prepare(SAVE_MOVIE) //preparamos la consulta
 	if err != nil {
 		return 0, err
@@ -145,7 +145,7 @@ func (r *repository) Save(ctx context.Context, m domain.Movie) (int64, error) {
 		return 0, err
 	}
 
-	return id, nil
+	return int(id), nil
 }
 
 func (r *repository) Update(ctx context.Context, m domain.Movie, id int) error {
